Initial commit of pathfinding system
Current implementation uses D*Lite with bias towards east, then south, then
west, then north. All cells have uniform cost and are only marked as blocked
(i.e., not traversable) if it is a ROCK or NONE tile.

The movement system was changed to go straight to the target location and do
nothing more.

In the future, merging the scan map should check to see if the world map needs
to grow to accommodate the changes, and tiles with rovers on them should be
marked as blocked only so long as they are visible.

The old movement strategy should be refactored in order to be used as a
fallback in case pathfinding is unneeded or cannot find a valid path.